tests/test_network_analysis.py
import unittest
from network_analysis import analyze_traffic, detect_intrusion, monitor_network
from utils.network_utils import parse_data, validate_input
from utils.logging_utils import log_action, log_error, log_info

class TestNetworkAnalysis(unittest.TestCase):

    def setUp(self):
        # Setup can include configuration of API keys, mock data, etc.
        self.mock_traffic_data = "mock_traffic_data_packet"
        self.expected_analysis_result = True
        self.expected_intrusion_result = False

    def test_analyze_traffic(self):
        # Test the analyze_traffic function from network_analysis.py
        try:
            result = analyze_traffic(self.mock_traffic_data)
            self.assertEqual(result, self.expected_analysis_result)
            log_info(f"Traffic analysis test passed with result: {result}")
        except Exception as e:
            log_error(f"Traffic analysis test failed with exception: {e}")
            self.fail(f"Traffic analysis test raised an exception: {e}")

    def test_detect_intrusion(self):
        # Test the detect_intrusion function from network_analysis.py
        try:
            result = detect_intrusion(self.mock_traffic_data)
            self.assertEqual(result, self.expected_intrusion_result)
            log_info(f"Intrusion detection test passed with result: {result}")
        except Exception as e:
            log_error(f"Intrusion detection test failed with exception: {e}")
            self.fail(f"Intrusion detection test raised an exception: {e}")

    def test_monitor_network(self):
        # Test the monitor_network function from network_analysis.py
        try:
            result = monitor_network(self.mock_traffic_data)
            self.assertTrue(validate_input(result))
            log_info(f"Network monitoring test passed with result: {result}")
        except Exception as e:
            log_error(f"Network monitoring test failed with exception: {e}")
            self.fail(f"Network monitoring test raised an exception: {e}")

    def test_parse_data(self):
        # Test the parse_data function from utils/network_utils.py
        try:
            result = parse_data(self.mock_traffic_data)
            self.assertIsNotNone(result)
            log_info(f"Data parsing test passed with result: {result}")
        except Exception as e:
            log_error(f"Data parsing test failed with exception: {e}")
            self.fail(f"Data parsing test raised an exception: {e}")

    def test_validate_input(self):
        # Test the validate_input function from utils/network_utils.py
        try:
            result = validate_input(self.mock_traffic_data)
            self.assertTrue(result)
            log_info(f"Input validation test passed with result: {result}")
        except Exception as e:
            log_error(f"Input validation test failed with exception: {e}")
            self.fail(f"Input validation test raised an exception: {e}")

if __name__ == '__main__':
    unittest.main()